Rename `BankInvoice.js` to `index.js`
